access control

+++ app/Http/Controllers/PostsController.php
@@ -11,6 +11,10 @@ use DB;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -93,6 +97,15 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error','Unauthorized page');
+        }
+
+        
+        
         return view('posts.edit')->with('post' , $post);
     }
 
@@ -129,8 +142,16 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
+
+        
         //
         $post=Post::find($id);
+        
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error','Unauthorized page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
